Update db; moved auth logic to its folder; update validate fucntion
