+++ apps/api/src/app.module.ts
@@ -1,11 +1,12 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import {PrismaService} from "./prisma.service";
-import {UserModule} from "./user/user.module";
+import { PrismaService } from './prisma.service';
+import { UserModule } from './user/user.module';
+import { AuthModule } from './auth/auth.module';
 
 @Module({
-  imports: [UserModule],
+  imports: [UserModule, AuthModule],
   controllers: [AppController],
   providers: [AppService, PrismaService],
 })
