--- apps/api/src/user/jwt.strategy.ts
@@ -1,25 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { PassportStrategy } from '@nestjs/passport';
-import { ExtractJwt, Strategy } from 'passport-jwt';
-
-interface JwtPayload {
-  id: string;
-  email: string;
-  role: string;
-  isSuspended: boolean;
-}
-
-@Injectable()
-export class JwtStrategy extends PassportStrategy(Strategy) {
-  constructor() {
-    super({
-      jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
-      ignoreExpiration: false,
-      secretOrKey: process.env.JWT_SECRET as string,
-    });
-  }
-
-  validate(payload: JwtPayload) {
-    return payload;
-  }
-}
